Refactor to move shared methods to BaseMatchService
